Subiendo casi todo andando, falta hacer el CRUD para la efi
