+++ app.py
@@ -1,9 +1,18 @@
 import os 
+import bcrypt
 from flask import Flask, render_template, redirect, request, url_for,jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from flask_jwt_extended import (
+    JWTManager,
+    )   
+from werkzeug.security import (
+    generate_password_hash,
+    check_password_hash,
+)
+from datetime import timedelta
 from dotenv import  load_dotenv
-
+from flask_marshmallow import Marshmallow
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('SQLALCHEMY_DATABASE_URI') # chequear a donde manda la informacion con respecto a la base de datos no encontrada de celulares.
@@ -11,6 +20,8 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
+jwt = JWTManager(app)
+ma = Marshmallow(app)
 
 from services.paises_service import PaisService
 from celulares import  *
@@ -19,287 +30,276 @@ from repositories.paises_repository import PaisRepository
 
 load_dotenv()
 
+from views import register_bp
+register_bp(app)
 
-#-----------------------EQUIPOS!!! (Index del programa) --------------------------
-@app.route('/', methods=['GET', 'POST'])
-def equipos():
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        modelo_id = request.form['modelo_id']
-        categoria_id = request.form['categoria_id']
-        costo = request.form['costo']
-        nuevo_equipo = Equipo(nombre=nombre,modelo_id = modelo_id, categoria_id=categoria_id, costo=costo)
-        db.session.add(nuevo_equipo)
-        db.session.commit()
-        return redirect(url_for('equipos'))
+
+# #-----------------------EQUIPOS!!! (Index del programa) --------------------------
+# @app.route('/', methods=['GET', 'POST'])
+# def equipos():
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         modelo_id = request.form['modelo_id']
+#         categoria_id = request.form['categoria_id']
+#         costo = request.form['costo']
+#         nuevo_equipo = Equipo(nombre=nombre,modelo_id = modelo_id, categoria_id=categoria_id, costo=costo)
+#         db.session.add(nuevo_equipo)
+#         db.session.commit()
+#         return redirect(url_for('equipos'))
     
-    equipos = Equipo.query.all()
-    modelos = Modelo.query.all()
-    categorias = Categoria.query.all()
-    return render_template('equipos.html', equipos=equipos, modelos=modelos, categorias=categorias)
-    
-    
-@app.route('/editar/<id>/equipos', methods=['GET', 'POST'])
-def editar_equipos(id):
-    equipo = Equipo.query.get_or_404(id)
-    modelos = Modelo.query.all()
-    categorias = Categoria.query.all()
-
-    if request.method == 'POST':
-        equipo.nombre = request.form['nombre']
-        equipo.modelo_id = request.form['modelo_id']
-        equipo.categoria_id = request.form['categoria_id']
-        equipo.costo = request.form['costo']
-        db.session.commit()
-        return redirect(url_for('equipos'))  # Redirige después de editar
-
-    return render_template('editar_equipos.html', equipo=equipo, modelos=modelos, categorias=categorias)
-
-
-
-#-----------------------PAISES--------------------------
-
-@app.route("/paises", methods=['POST', 'GET'])
-def paises(): 
-    formulario = PaisForm()
-
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        services = PaisService(PaisRepository())
-        paises = services.get_all()
-        services.create(nombre=nombre)    
-    services = PaisService(PaisRepository())
-    paises = services.get_all()
-    return render_template(
-        'paises.html',
-        paises = paises,
-        formulario = formulario)
-
-
-@app.route('/editar/<id>/paises', methods=['GET', 'POST'])
-def editar_paises(id):
-    pais = Pais.query.get_or_404(id)
-    if request.method == 'POST':
-        pais.nombre = request.form['nombre']
-        db.session.commit()
-        return redirect(url_for('paises'))  # Redirige después de editar
-
-    return render_template('editar_paises.html', pais=pais)
-
-
-
-#-----------------------MODELOS--------------------------
-@app.route('/modelos', methods=['GET', 'POST'])
-def modelos():
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        fabricante_id = request.form['fabricante_id']
-        nuevo_modelo = Modelo(nombre=nombre, fabricante_id=fabricante_id)
-        db.session.add(nuevo_modelo)
-        db.session.commit()
-        return redirect(url_for('modelos'))
-    
-    modelos = Modelo.query.all()
-    fabricantes = Fabricante.query.all()
-    return render_template('modelos.html', modelos=modelos, fabricantes=fabricantes)
+#     equipos = Equipo.query.all()
+#     modelos = Modelo.query.all()
+#     categorias = Categoria.query.all()
+#     return render_template('equipos.html', equipos=equipos, modelos=modelos, categorias=categorias)
     
     
-@app.route('/editar/<id>/modelos', methods=['GET', 'POST'])
-def editar_modelo(id):
-    modelo = Modelo.query.get_or_404(id)
-    fabricantes = Fabricante.query.all()
+# @app.route('/editar/<id>/equipos', methods=['GET', 'POST'])
+# def editar_equipos(id):
+#     equipo = Equipo.query.get_or_404(id)
+#     modelos = Modelo.query.all()
+#     categorias = Categoria.query.all()
 
-    if request.method == 'POST':
-        modelo.nombre = request.form['nombre']
-        modelo.modelo_id = request.form['fabricante_id']
-        db.session.commit()
-        return redirect(url_for('modelos'))  # Redirige después de editar
+#     if request.method == 'POST':
+#         equipo.nombre = request.form['nombre']
+#         equipo.modelo_id = request.form['modelo_id']
+#         equipo.categoria_id = request.form['categoria_id']
+#         equipo.costo = request.form['costo']
+#         db.session.commit()
+#         return redirect(url_for('equipos'))  # Redirige después de editar
 
-    return render_template('editar_modelos.html', modelo=modelo, fabricantes=fabricantes)
+#     return render_template('editar_equipos.html', equipo=equipo, modelos=modelos, categorias=categorias)
 
-#-----------------------CATEGORIAS--------------------------
 
-@app.route("/categorias", methods=['POST', 'GET'])
-def categorias(): 
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        nueva_categoria = Categoria(nombre=nombre)
-        db.session.add(nueva_categoria)
-        db.session.commit()
-    categorias_query = Categoria.query.all()
-    return render_template('categorias.html',categorias = categorias_query)
 
-@app.route('/editar/<id>/categorias', methods=['GET', 'POST'])
-def editar_categorias(id):
-    categoria = Categoria.query.get_or_404(id)
-    if request.method == 'POST':
-        categoria.nombre = request.form['nombre']
-        db.session.commit()
-        return redirect(url_for('categorias'))  # Redirige después de editar
+# #-----------------------PAISES--------------------------
 
-    return render_template('editar_categorias.html', categoria = categoria)
+# @app.route("/paises", methods=['POST', 'GET'])
+# def paises(): 
+#     formulario = PaisForm()
+
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         services = PaisService(PaisRepository())
+#         paises = services.get_all()
+#         services.create(nombre=nombre)    
+#     services = PaisService(PaisRepository())
+#     paises = services.get_all()
+#     return render_template(
+#         'paises.html',
+#         paises = paises,
+#         formulario = formulario)
+
+
+# @app.route('/editar/<id>/paises', methods=['GET', 'POST'])
+# def editar_paises(id):
+#     pais = Pais.query.get_or_404(id)
+#     if request.method == 'POST':
+#         pais.nombre = request.form['nombre']
+#         db.session.commit()
+#         return redirect(url_for('paises'))  # Redirige después de editar
+
+#     return render_template('editar_paises.html', pais=pais)
+
+
+
+# #-----------------------MODELOS--------------------------
+# @app.route('/modelos', methods=['GET', 'POST'])
+# def modelos():
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         fabricante_id = request.form['fabricante_id']
+#         nuevo_modelo = Modelo(nombre=nombre, fabricante_id=fabricante_id)
+#         db.session.add(nuevo_modelo)
+#         db.session.commit()
+#         return redirect(url_for('modelos'))
     
-#------------------------ACCESORIOS-----------------------
-@app.route("/accesorios", methods=['POST', 'GET'])
-def accesorios(): 
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        nuevo_accesorio = Accesorio(nombre=nombre)
-        db.session.add(nuevo_accesorio)
-        db.session.commit()
-    accesorios_query = Accesorio.query.all()
-    return render_template('accesorios.html',accesorios = accesorios_query)
-
-@app.route('/editar/<id>/accesorios', methods=['GET', 'POST'])
-def editar_accesorio(id):
-    accesorio = Accesorio.query.get_or_404(id)
+#     modelos = Modelo.query.all()
+#     fabricantes = Fabricante.query.all()
+#     return render_template('modelos.html', modelos=modelos, fabricantes=fabricantes)
     
-    if request.method == 'POST':
-        accesorio.nombre = request.form['nombre']
-        db.session.commit()
-        return redirect(url_for('accesorios'))  # Redirige después de editar
-
-    return render_template('editar_accesorios.html', accesorio = accesorio)
-
-#------------------------CARACTERISTICAS-----------------------
-@app.route("/caracteristicas", methods=['POST', 'GET'])
-def caracteristicas(): 
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        descripcion = request.form['descripcion']
-        nueva_caracteristica = Caracteristica(nombre=nombre,descripcion=descripcion)
-        db.session.add(nueva_caracteristica)
-        db.session.commit()
-        return redirect(url_for('caracteristicas'))  # Redirige después de editar
-    caracteristicas = Caracteristica.query.all()
-    return render_template('caracteristicas.html', caracteristicas = caracteristicas)
-
-@app.route('/editar/<id>/caracteristicas', methods=['GET', 'POST'])
-def editar_caracteristica(id):
-    caracteristica = Caracteristica.query.get_or_404(id)
     
-    if request.method == 'POST':
-        caracteristica.nombre = request.form['nombre']
-        caracteristica.descripcion = request.form['descripcion']
-        db.session.commit()
-        return redirect(url_for('caracteristicas'))  # Redirige después de editar
+# @app.route('/editar/<id>/modelos', methods=['GET', 'POST'])
+# def editar_modelo(id):
+#     modelo = Modelo.query.get_or_404(id)
+#     fabricantes = Fabricante.query.all()
 
-    return render_template('editar_caracteristicas.html', caracteristica=caracteristica)
+#     if request.method == 'POST':
+#         modelo.nombre = request.form['nombre']
+#         modelo.modelo_id = request.form['fabricante_id']
+#         db.session.commit()
+#         return redirect(url_for('modelos'))  # Redirige después de editar
 
-#------------------------CARACTERISTICAS-MODELOS-----------------------
-@app.route('/caracteristicas_modelos', methods=['GET', 'POST'])
-def caract_model():
-    if request.method == 'POST':
-        caracteristica_id = request.form['caracteristica_id']
-        modelo_id = request.form['modelo_id']
-        nuevo_caract_model = CaracteristicaModelo(modelo_id=modelo_id, caracteristica_id=caracteristica_id)
-        db.session.add(nuevo_caract_model)
-        db.session.commit()
-        return redirect(url_for('caract_model'))
+#     return render_template('editar_modelos.html', modelo=modelo, fabricantes=fabricantes)
 
-    caracts_models = CaracteristicaModelo.query.all()
-    caracteristicas = Caracteristica.query.all()  # Traer todas las características
-    modelos = Modelo.query.all()  # Traer todos los modelos
+# #-----------------------CATEGORIAS--------------------------
+
+# @app.route("/categorias", methods=['POST', 'GET'])
+# def categorias(): 
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         nueva_categoria = Categoria(nombre=nombre)
+#         db.session.add(nueva_categoria)
+#         db.session.commit()
+#     categorias_query = Categoria.query.all()
+#     return render_template('categorias.html',categorias = categorias_query)
+
+# @app.route('/editar/<id>/categorias', methods=['GET', 'POST'])
+# def editar_categorias(id):
+#     categoria = Categoria.query.get_or_404(id)
+#     if request.method == 'POST':
+#         categoria.nombre = request.form['nombre']
+#         db.session.commit()
+#         return redirect(url_for('categorias'))  # Redirige después de editar
+
+#     return render_template('editar_categorias.html', categoria = categoria)
     
-    return render_template('caracteristicas_modelos.html', caracts_models=caracts_models, caracteristicas=caracteristicas, modelos=modelos)
+# #------------------------ACCESORIOS-----------------------
+# @app.route("/accesorios", methods=['POST', 'GET'])
+# def accesorios(): 
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         nuevo_accesorio = Accesorio(nombre=nombre)
+#         db.session.add(nuevo_accesorio)
+#         db.session.commit()
+#     accesorios_query = Accesorio.query.all()
+#     return render_template('accesorios.html',accesorios = accesorios_query)
 
-#------------------------ACCESORIOS-MODELOS-----------------------
-@app.route('/accesorios_modelos', methods=['GET', 'POST'])
-def acces_model():
-    if request.method == 'POST':
-        accesorio_id = request.form['accesorio_id']
-        modelo_id = request.form['modelo_id']
-        nuevo_acces_model = AccesorioModelo(accesorio_id = accesorio_id, modelo_id = modelo_id)
-        db.session.add(nuevo_acces_model)
-        db.session.commit()
-        return redirect(url_for('acces_model'))
-
-    acces_models = AccesorioModelo.query.all()
-    accesorios = Accesorio.query.all()  # Traer todos los accesorios
-    modelos = Modelo.query.all()  # Traer todos los modelos
+# @app.route('/editar/<id>/accesorios', methods=['GET', 'POST'])
+# def editar_accesorio(id):
+#     accesorio = Accesorio.query.get_or_404(id)
     
-    return render_template('accesorios_modelos.html', acces_models=acces_models, accesorios=accesorios, modelos=modelos)
+#     if request.method == 'POST':
+#         accesorio.nombre = request.form['nombre']
+#         db.session.commit()
+#         return redirect(url_for('accesorios'))  # Redirige después de editar
 
-@app.route('/editar/<id>/accesorios_modelos', methods=['GET', 'POST'])
-def editar_acces_model(id):
-    accesorio = AccesorioModelo.query.get_or_404(id)
-    accesorios = Accesorio.query.all() 
-    modelos = Modelo.query.all()
+#     return render_template('editar_accesorios.html', accesorio = accesorio)
+
+# #------------------------CARACTERISTICAS-----------------------
+# @app.route("/caracteristicas", methods=['POST', 'GET'])
+# def caracteristicas(): 
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         descripcion = request.form['descripcion']
+#         nueva_caracteristica = Caracteristica(nombre=nombre,descripcion=descripcion)
+#         db.session.add(nueva_caracteristica)
+#         db.session.commit()
+#         return redirect(url_for('caracteristicas'))  # Redirige después de editar
+#     caracteristicas = Caracteristica.query.all()
+#     return render_template('caracteristicas.html', caracteristicas = caracteristicas)
+
+# @app.route('/editar/<id>/caracteristicas', methods=['GET', 'POST'])
+# def editar_caracteristica(id):
+#     caracteristica = Caracteristica.query.get_or_404(id)
     
-    if request.method == 'POST':
-        accesorio.accesorio_id = request.form['accesorio_id']
-        accesorio.modelo_id = request.form['modelo_id']
-        db.session.commit()
-        return redirect(url_for('acces_model'))  # Redirige después de editar
+#     if request.method == 'POST':
+#         caracteristica.nombre = request.form['nombre']
+#         caracteristica.descripcion = request.form['descripcion']
+#         db.session.commit()
+#         return redirect(url_for('caracteristicas'))  # Redirige después de editar
 
-    return render_template('editar_acces_mod.html', accesorio=accesorio, accesorios=accesorios, modelos=modelos)
+#     return render_template('editar_caracteristicas.html', caracteristica=caracteristica)
 
-#-----------------------FABRICANTES-----------------------
-@app.route('/fabricantes', methods=['GET', 'POST'])
-def fabricantes():
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        pais_id = request.form['pais_id']
-        nuevo_fabricante = Fabricante(nombre=nombre, pais_id=pais_id)
-        db.session.add(nuevo_fabricante)
-        db.session.commit()
-        return redirect(url_for('fabricantes'))  # Redirige para evitar el duplicado en caso de recarga
+# #------------------------CARACTERISTICAS-MODELOS-----------------------
+# @app.route('/caracteristicas_modelos', methods=['GET', 'POST'])
+# def caract_model():
+#     if request.method == 'POST':
+#         caracteristica_id = request.form['caracteristica_id']
+#         modelo_id = request.form['modelo_id']
+#         nuevo_caract_model = CaracteristicaModelo(modelo_id=modelo_id, caracteristica_id=caracteristica_id)
+#         db.session.add(nuevo_caract_model)
+#         db.session.commit()
+#         return redirect(url_for('caract_model'))
 
-    fabricantes = Fabricante.query.all()
-    paises = Pais.query.all()
-    return render_template('fabricantes.html', fabricantes=fabricantes, paises=paises)
-
-@app.route('/editar/<id>/fabricantes', methods=['GET', 'POST'])
-def editar_fabricante(id):
-    fabricante = Fabricante.query.get_or_404(id)
-    paises = Pais.query.all()
+#     caracts_models = CaracteristicaModelo.query.all()
+#     caracteristicas = Caracteristica.query.all()  # Traer todas las características
+#     modelos = Modelo.query.all()  # Traer todos los modelos
     
-    if request.method == 'POST':
-        fabricante.nombre = request.form['nombre']
-        fabricante.pais_id = request.form['pais_id']
-        db.session.commit()
-        return redirect(url_for('fabricantes'))  # Redirige después de editar
+#     return render_template('caracteristicas_modelos.html', caracts_models=caracts_models, caracteristicas=caracteristicas, modelos=modelos)
 
-    return render_template('editar_fabricantes.html', fabricante=fabricante, paises=paises)
+# #------------------------ACCESORIOS-MODELOS-----------------------
+# @app.route('/accesorios_modelos', methods=['GET', 'POST'])
+# def acces_model():
+#     if request.method == 'POST':
+#         accesorio_id = request.form['accesorio_id']
+#         modelo_id = request.form['modelo_id']
+#         nuevo_acces_model = AccesorioModelo(accesorio_id = accesorio_id, modelo_id = modelo_id)
+#         db.session.add(nuevo_acces_model)
+#         db.session.commit()
+#         return redirect(url_for('acces_model'))
 
-#------------------------PROVEEDORES-----------------------
-@app.route("/proveedores", methods=['POST', 'GET'])
-def proveedores(): 
-    if request.method == 'POST':
-        nombre = request.form['nombre']
-        cuit = request.form['cuit']
-        nuevo_proveedor = Proveedor(nombre=nombre,cuit=cuit)
-        db.session.add(nuevo_proveedor)
-        db.session.commit()
-        return redirect(url_for('proveedores'))  # Redirige después de editar
-    proveedores = Proveedor.query.all()
-    return render_template('proveedores.html', proveedores = proveedores)
-
-@app.route('/editar/<id>/proveedores', methods=['GET', 'POST'])
-def editar_proveedor(id):
-    proveedor = Proveedor.query.get_or_404(id)
+#     acces_models = AccesorioModelo.query.all()
+#     accesorios = Accesorio.query.all()  # Traer todos los accesorios
+#     modelos = Modelo.query.all()  # Traer todos los modelos
     
-    if request.method == 'POST':
-        proveedor.nombre = request.form['nombre']
-        proveedor.cuit= request.form['cuit']
-        db.session.commit()
-        return redirect(url_for('proveedores'))  # Redirige después de editar
+#     return render_template('accesorios_modelos.html', acces_models=acces_models, accesorios=accesorios, modelos=modelos)
 
-    return render_template('editar_proveedores.html', proveedor=proveedor)
-
-
-
-
-@app.route("/users", methods=['POST'])
-def user(): 
-    data = request.get_json()
-    nombre = data.get('nombre')
-    password = data.get('password')
-    try:
-        nuevo_usuario = Usuario(nombre=nombre, password=password)
-        db.session.add(nuevo_usuario)
-        db.session.commit()
-        return jsonify({'Usuario Creado':nombre}),201    
-    except:
-        return jsonify({'Mensaje': 'Algo salio mal'})    
+# @app.route('/editar/<id>/accesorios_modelos', methods=['GET', 'POST'])
+# def editar_acces_model(id):
+#     accesorio = AccesorioModelo.query.get_or_404(id)
+#     accesorios = Accesorio.query.all() 
+#     modelos = Modelo.query.all()
     
+#     if request.method == 'POST':
+#         accesorio.accesorio_id = request.form['accesorio_id']
+#         accesorio.modelo_id = request.form['modelo_id']
+#         db.session.commit()
+#         return redirect(url_for('acces_model'))  # Redirige después de editar
+
+#     return render_template('editar_acces_mod.html', accesorio=accesorio, accesorios=accesorios, modelos=modelos)
+
+# #-----------------------FABRICANTES-----------------------
+# @app.route('/fabricantes', methods=['GET', 'POST'])
+# def fabricantes():
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         pais_id = request.form['pais_id']
+#         nuevo_fabricante = Fabricante(nombre=nombre, pais_id=pais_id)
+#         db.session.add(nuevo_fabricante)
+#         db.session.commit()
+#         return redirect(url_for('fabricantes'))  # Redirige para evitar el duplicado en caso de recarga
+
+#     fabricantes = Fabricante.query.all()
+#     paises = Pais.query.all()
+#     return render_template('fabricantes.html', fabricantes=fabricantes, paises=paises)
+
+# @app.route('/editar/<id>/fabricantes', methods=['GET', 'POST'])
+# def editar_fabricante(id):
+#     fabricante = Fabricante.query.get_or_404(id)
+#     paises = Pais.query.all()
+    
+#     if request.method == 'POST':
+#         fabricante.nombre = request.form['nombre']
+#         fabricante.pais_id = request.form['pais_id']
+#         db.session.commit()
+#         return redirect(url_for('fabricantes'))  # Redirige después de editar
+
+#     return render_template('editar_fabricantes.html', fabricante=fabricante, paises=paises)
+
+# #------------------------PROVEEDORES-----------------------
+# @app.route("/proveedores", methods=['POST', 'GET'])
+# def proveedores(): 
+#     if request.method == 'POST':
+#         nombre = request.form['nombre']
+#         cuit = request.form['cuit']
+#         nuevo_proveedor = Proveedor(nombre=nombre,cuit=cuit)
+#         db.session.add(nuevo_proveedor)
+#         db.session.commit()
+#         return redirect(url_for('proveedores'))  # Redirige después de editar
+#     proveedores = Proveedor.query.all()
+#     return render_template('proveedores.html', proveedores = proveedores)
+
+# @app.route('/editar/<id>/proveedores', methods=['GET', 'POST'])
+# def editar_proveedor(id):
+#     proveedor = Proveedor.query.get_or_404(id)
+    
+#     if request.method == 'POST':
+#         proveedor.nombre = request.form['nombre']
+#         proveedor.cuit= request.form['cuit']
+#         db.session.commit()
+#         return redirect(url_for('proveedores'))  # Redirige después de editar
+
+#     return render_template('editar_proveedores.html', proveedor=proveedor)
+
+
+
