+++ docker-compose.yaml
@@ -1,0 +1,27 @@
+
+# sudo docker build --tag first_docker_b .
+services: # Lista de servicios
+  my-app-name: #servicio/app de flask
+    build: # Lista de parámetros desde los cuales construir
+      context: . #contexto
+      dockerfile: Dockerfile #archivo de Dockerfile
+
+
+# sudo docker build --tag first_docker_b .
+# es reemplazado por docker-compose build
+
+# docker run -d -p 5005:5005 first_docker_b
+# es reemplazado por docker-compose up
+    ports: #puerto contenedor:puerto host
+      - "5005:5005"
+    restart: always #en caso de error o falla
+#  mysql_db:
+
+
+#BUILDEO DE CONTENEDOR
+# sudo docker build --tag first_docker_b .
+
+#CORRER EL CONTENDOR
+#- sudo docker run -d -p 5005:5005 first_docker_b mysql:5.7 -u root -pw root -p 3600 -e MYSQL_ROOT_PASSWORD=rootpassword -e 
+
+# Se reemplaza por docker-compose up y docker-compose up -d
