+++ requirements.txt
@@ -1,7 +1,11 @@
 alembic==1.13.2
+bcrypt==4.2.0
 blinker==1.8.2
 click==8.1.7
 Flask==3.0.3
+Flask-JWT==0.3.2
+Flask-JWT-Extended==4.6.0
+flask-marshmallow==1.2.1
 Flask-Migrate==4.0.7
 Flask-SQLAlchemy==3.1.1
 Flask-WTF==1.2.1
@@ -10,7 +14,12 @@ itsdangerous==2.2.0
 Jinja2==3.1.4
 Mako==1.3.5
 MarkupSafe==2.1.5
+marshmallow==3.22.0
+marshmallow-sqlalchemy==1.1.0
+packaging==24.1
+PyJWT==2.9.0
 PyMySQL==1.1.1
+python-dotenv==1.0.1
 SQLAlchemy==2.0.31
 typing_extensions==4.12.2
 Werkzeug==3.0.3
